Restructured Firebase Hierchy
It didn't really make any sense how I had it before. Now, which is 100% the easiest way, each user will be assigned a unique userID which will be stored in the database. That way, the authentification can just get the current user based on username and password then store that key to access the data. This could also be very easily expandable with google, facebook, phone, etc. sign ins. That would be pretty sick and would probably only be another 2-3 hours worth of work for implementation, but can be worried about later if time permits.